# Conflicts:
#	pxr/base/lib/arch/attributes.h
#	pxr/base/lib/arch/demangle.h
#	pxr/base/lib/arch/errno.h
#	pxr/base/lib/arch/error.h
#	pxr/base/lib/arch/fileSystem.cpp
#	pxr/base/lib/arch/fileSystem.h
#	pxr/base/lib/arch/function.h
#	pxr/base/lib/arch/hash.h
#	pxr/base/lib/arch/mallocHook.h
#	pxr/base/lib/arch/math.h
#	pxr/base/lib/arch/nap.h
#	pxr/base/lib/arch/stackTrace.h
#	pxr/base/lib/arch/symbols.h
#	pxr/base/lib/arch/systemInfo.h
#	pxr/base/lib/arch/testenv/testFileSystem.cpp
#	pxr/base/lib/arch/threads.h
#	pxr/base/lib/arch/timing.h
#	pxr/base/lib/arch/vsnprintf.h
#	pxr/base/lib/gf/bbox3d.h
#	pxr/base/lib/gf/colorRamp.h
#	pxr/base/lib/gf/frustum.h
#	pxr/base/lib/gf/gamma.h
#	pxr/base/lib/gf/homogeneous.h
#	pxr/base/lib/gf/interval.h
#	pxr/base/lib/gf/line.h
#	pxr/base/lib/gf/line2d.h
#	pxr/base/lib/gf/lineSeg.h
#	pxr/base/lib/gf/lineSeg2d.h
#	pxr/base/lib/gf/math.h
#	pxr/base/lib/gf/matrix2d.h
#	pxr/base/lib/gf/matrix2f.h
#	pxr/base/lib/gf/matrix3d.h
#	pxr/base/lib/gf/matrix3f.h
#	pxr/base/lib/gf/matrix4d.h
#	pxr/base/lib/gf/matrix4f.h
#	pxr/base/lib/gf/matrixData.h
#	pxr/base/lib/gf/plane.h
#	pxr/base/lib/gf/quatd.h
#	pxr/base/lib/gf/quaternion.h
#	pxr/base/lib/gf/quatf.h
#	pxr/base/lib/gf/quath.h
#	pxr/base/lib/gf/range1d.h
#	pxr/base/lib/gf/range1f.h
#	pxr/base/lib/gf/range2d.h
#	pxr/base/lib/gf/range2f.h
#	pxr/base/lib/gf/range3d.h
#	pxr/base/lib/gf/range3f.h
#	pxr/base/lib/gf/ray.h
#	pxr/base/lib/gf/rect2i.h
#	pxr/base/lib/gf/rgb.h
#	pxr/base/lib/gf/rgba.h
#	pxr/base/lib/gf/rotation.h
#	pxr/base/lib/gf/size2.h
#	pxr/base/lib/gf/size3.h
#	pxr/base/lib/gf/transform.h
#	pxr/base/lib/gf/vec3d.h
#	pxr/base/lib/gf/vec3f.h
#	pxr/base/lib/gf/vec3h.h
#	pxr/base/lib/js/json.h
#	pxr/base/lib/js/value.h
#	pxr/base/lib/plug/plugin.h
#	pxr/base/lib/tf/anyWeakPtr.h
#	pxr/base/lib/tf/atomicOfstreamWrapper.h
#	pxr/base/lib/tf/callContext.h
#	pxr/base/lib/tf/debug.h
#	pxr/base/lib/tf/diagnosticLite.h
#	pxr/base/lib/tf/diagnosticMgr.h
#	pxr/base/lib/tf/diagnosticNotice.h
#	pxr/base/lib/tf/dl.h
#	pxr/base/lib/tf/enum.h
#	pxr/base/lib/tf/envSetting.h
#	pxr/base/lib/tf/errorMark.h
#	pxr/base/lib/tf/errorTransport.h
#	pxr/base/lib/tf/expiryNotifier.h
#	pxr/base/lib/tf/fileUtils.h
#	pxr/base/lib/tf/getenv.h
#	pxr/base/lib/tf/hash.h
#	pxr/base/lib/tf/instantiateSingleton.h
#	pxr/base/lib/tf/iterator.h
#	pxr/base/lib/tf/mallocTag.h
#	pxr/base/lib/tf/notice.h
#	pxr/base/lib/tf/noticeRegistry.h
#	pxr/base/lib/tf/nullPtr.h
#	pxr/base/lib/tf/pathUtils.h
#	pxr/base/lib/tf/patternMatcher.h
#	pxr/base/lib/tf/pointerAndBits.h
#	pxr/base/lib/tf/poolAllocator.h
#	pxr/base/lib/tf/preprocessorUtils.h
#	pxr/base/lib/tf/pyEnum.h
#	pxr/base/lib/tf/pyInterpreter.h
#	pxr/base/lib/tf/pyLock.h
#	pxr/base/lib/tf/pyObjWrapper.h
#	pxr/base/lib/tf/pyObjectFinder.h
#	pxr/base/lib/tf/pyTracing.h
#	pxr/base/lib/tf/pyUtils.h
#	pxr/base/lib/tf/pyWeakObject.h
#	pxr/base/lib/tf/pyWrapContext.h
#	pxr/base/lib/tf/refBase.h
#	pxr/base/lib/tf/refPtr.h
#	pxr/base/lib/tf/refPtrTracker.h
#	pxr/base/lib/tf/regTest.h
#	pxr/base/lib/tf/registryManager.h
#	pxr/base/lib/tf/scopeDescription.h
#	pxr/base/lib/tf/scriptModuleLoader.h
#	pxr/base/lib/tf/singleton.h
#	pxr/base/lib/tf/staticTokens.h
#	pxr/base/lib/tf/stl.h
#	pxr/base/lib/tf/stopwatch.h
#	pxr/base/lib/tf/stringUtils.h
#	pxr/base/lib/tf/token.h
#	pxr/base/lib/tf/type.h
#	pxr/base/lib/tf/weakBase.h
#	pxr/base/lib/tf/wrapTypeHelpers.h
#	pxr/base/lib/tracelite/trace.h
#	pxr/base/lib/vt/array.h
#	pxr/base/lib/vt/dictionary.h
#	pxr/base/lib/vt/value.h
#	pxr/base/lib/work/detachedTask.h
#	pxr/base/lib/work/threadLimits.h
#	pxr/usd/lib/usdShade/look.h
#	third_party/maya/lib/pxrUsdMayaGL/proxyDrawOverride.h

+++ pxr/base/lib/gf/size3.h
@@ -24,59 +24,55 @@
 #ifndef GF_SIZE3_H
 #define GF_SIZE3_H
 
+/// \file gf/size3.h
+/// \ingroup group_gf_LinearAlgebra
+
 #include "pxr/base/arch/inttypes.h"
 #include "pxr/base/gf/vec3i.h"
 #include "pxr/base/gf/api.h" 
 
 #include <iosfwd>
 
-/*!
- * \file size3.h
- * \ingroup group_gf_LinearAlgebra
- */
-
-/*!
- * \class GfSize3 size3.h pxr/base/gf/size3.h
- * \ingroup group_gf_LinearAlgebra
- * \brief Three-dimensional array of sizes
- *
- * GfSize3 is used to represent triples of counts.  It is based on
- * the datatype size_t, and thus can only represent non-negative
- * values in each dimension.  If you need to represent negative numbers
- * as well, use GfVeci.
- *
- * Usage of GfSize3 is similar to that of GfVec3i, except that all
- * mathematical operations are componentwise (including multiplication).
- */
-
+/// \class GfSize3
+/// \ingroup group_gf_LinearAlgebra
+///
+/// Three-dimensional array of sizes
+///
+/// GfSize3 is used to represent triples of counts.  It is based on the
+/// datatype size_t, and thus can only represent non-negative values in each
+/// dimension.  If you need to represent negative numbers as well, use GfVeci.
+///
+/// Usage of GfSize3 is similar to that of GfVec3i, except that all
+/// mathematical operations are componentwise (including multiplication).
+///
 class GfSize3 {
 public:
-    //! Default constructor initializes components to zero
+    /// Default constructor initializes components to zero
     GfSize3() {
-	Set(0, 0, 0);
+        Set(0, 0, 0);
     }
 
-    //! Copy constructor.
+    /// Copy constructor.
     GfSize3(const GfSize3& o) {
         *this = o;
     }
 
-    //! Conversion from GfVec3i
+    /// Conversion from GfVec3i
     explicit GfSize3(const GfVec3i&o) {
-	Set(o[0], o[1], o[2]);
+        Set(o[0], o[1], o[2]);
     }
 
-    //! Construct from an array
+    /// Construct from an array
     GfSize3(const size_t v[3]) {
         Set(v);
     }
 
-    //! Construct from three values
+    /// Construct from three values
     GfSize3(size_t v0, size_t v1, size_t v2) {
         Set(v0, v1, v2);
     }
 
-    //! Set to the values in \p v.
+    /// Set to the values in \p v.
     GfSize3 & Set(const size_t v[3]) {
         _vec[0] = v[0]; 
         _vec[1] = v[1]; 
@@ -84,7 +80,7 @@ public:
         return *this;
     }
 
-    //! Set to values passed directly
+    /// Set to values passed directly
     GfSize3 & Set(size_t v0, size_t v1, size_t v2) {
         _vec[0] = v0; 
         _vec[1] = v1; 
@@ -92,28 +88,28 @@ public:
         return *this;
     }
 
-    //! Array operator
+    /// Array operator
     size_t & operator [](size_t i) {
         return _vec[i];
     }
 
-    //! Const array operator
+    /// Const array operator
     const size_t & operator [](size_t i) const {
         return _vec[i];
     }
 
-    //! Component-wise equality
+    /// Component-wise equality
     bool operator ==(const GfSize3 &v) const {
         return _vec[0] == v._vec[0] && _vec[1] == v._vec[1] &&
-	    _vec[2] == v._vec[2];
+            _vec[2] == v._vec[2];
     }
 
-    //! Component-wise inequality
+    /// Component-wise inequality
     bool operator !=(const GfSize3 &v) const {
         return ! (*this == v);
     }
 
-    //! Component-wise in-place addition
+    /// Component-wise in-place addition
     GfSize3 & operator +=(const GfSize3 &v) {
         _vec[0] += v._vec[0]; 
         _vec[1] += v._vec[1]; 
@@ -121,7 +117,7 @@ public:
         return *this;
     }
 
-    //! Component-wise in-place subtraction
+    /// Component-wise in-place subtraction
     GfSize3 & operator -=(const GfSize3 &v) {
         _vec[0] -= v._vec[0]; 
         _vec[1] -= v._vec[1]; 
@@ -129,7 +125,7 @@ public:
         return *this;
     }
 
-    //! Component-wise in-place multiplication.
+    /// Component-wise in-place multiplication.
     GfSize3 & operator *=(GfSize3 const &v) {
         _vec[0] *= v._vec[0];
         _vec[1] *= v._vec[1];
@@ -137,7 +133,7 @@ public:
         return *this;
     }
 
-    //! Component-wise in-place multiplication by a scalar
+    /// Component-wise in-place multiplication by a scalar
     GfSize3 & operator *=(size_t d) {
         _vec[0] = _vec[0] * d;
         _vec[1] = _vec[1] * d;
@@ -145,7 +141,7 @@ public:
         return *this;
     }
 
-    //! Component-wise in-place division by a scalar
+    /// Component-wise in-place division by a scalar
     GfSize3 & operator /=(size_t d) {
         _vec[0] = _vec[0] / d;
         _vec[1] = _vec[1] / d;
@@ -153,54 +149,54 @@ public:
         return *this;
     }
 
-    //! Component-wise addition
+    /// Component-wise addition
     friend GfSize3 operator +(const GfSize3 &v1, const GfSize3 &v3) {
         return GfSize3(v1._vec[0]+v3._vec[0],
-		       v1._vec[1]+v3._vec[1],
-		       v1._vec[2]+v3._vec[2]);
+                       v1._vec[1]+v3._vec[1],
+                       v1._vec[2]+v3._vec[2]);
     }
 
-    //! Component-wise subtraction
+    /// Component-wise subtraction
     friend GfSize3 operator -(const GfSize3 &v1, const GfSize3 &v3) {
         return GfSize3(v1._vec[0]-v3._vec[0],
-		       v1._vec[1]-v3._vec[1],
-		       v1._vec[2]-v3._vec[2]);
+                       v1._vec[1]-v3._vec[1],
+                       v1._vec[2]-v3._vec[2]);
     }
 
-    //! Component-wise multiplication
+    /// Component-wise multiplication
     friend GfSize3 operator *(const GfSize3 &v1, const GfSize3 &v3) {
         return GfSize3(v1._vec[0]*v3._vec[0],
-		       v1._vec[1]*v3._vec[1],
-		       v1._vec[2]*v3._vec[2]);
+                       v1._vec[1]*v3._vec[1],
+                       v1._vec[2]*v3._vec[2]);
     }
 
-    //! Component-wise multiplication by a scalar
+    /// Component-wise multiplication by a scalar
     friend GfSize3 operator *(const GfSize3 &v1, size_t s) {
         return GfSize3(v1._vec[0]*s,
-		       v1._vec[1]*s,
-		       v1._vec[2]*s);
+                       v1._vec[1]*s,
+                       v1._vec[2]*s);
     }
 
-    //! Component-wise multiplication by a scalar
+    /// Component-wise multiplication by a scalar
     friend GfSize3 operator *(size_t s, const GfSize3 &v1) {
         return GfSize3(v1._vec[0]*s,
-		       v1._vec[1]*s,
-		       v1._vec[2]*s);
+                       v1._vec[1]*s,
+                       v1._vec[2]*s);
     }
 
-    //! Component-wise division by a scalar
+    /// Component-wise division by a scalar
     friend GfSize3 operator /(const GfSize3 &v1, size_t s) {
         return GfSize3(v1._vec[0]/s,
-		       v1._vec[1]/s,
-		       v1._vec[2]/s);
+                       v1._vec[1]/s,
+                       v1._vec[2]/s);
     }
 
-    //! Output operator
+    /// Output operator
     friend GF_API std::ostream &operator<<(std::ostream &o, GfSize3 const &v);
 
-    //! Conversion to GfVec3i
+    /// Conversion to GfVec3i
     operator GfVec3i() const {
-	return GfVec3i(static_cast<int>(_vec[0]), static_cast<int>(_vec[1]), static_cast<int>(_vec[2]));
+        return GfVec3i(_vec[0],_vec[1],_vec[2]);
     }
 private:
     size_t _vec[3];
